Communicate file extension to Server
Close #24

As described in #24 the file extension had no impact on the behavior.
Even without specifying it, the rendered videos inside the feedback
client window appeared to work just as expected, however, in order
to make the project bulletproof for the future it is better to
extract the file encoding and communicate it correctly.